requirements.txt
Django
djangorestframework
django-cors-headers
django-filter==21.1
